+++ ConanProfiles/MacOS-clang-Release.ini
@@ -1,0 +1,13 @@
+[settings]
+arch=armv8
+os=Macos
+build_type=Release
+compiler=apple-clang
+compiler.version=14
+compiler.libcxx=libstdc++
+[options]
+[build_requires]
+[env]
+CXX=clang++
+CC=clang
+CXXFLAGS="-O3 -Wall"
