+++ src/Interfaces/Full/MotorController.cpp
@@ -82,3 +82,55 @@ void MotorController::newFrame(const can_frame& frame) {
 void MotorController::newError() {
     fmt::print("Error\n");
 };
+
+enum Outputs {
+    // 0x0A0 - Temps #1
+    ModuleATemp,
+    ModuleBTemp,
+    ModuleCTemp,
+    GateDriverTemp,
+
+    // 0x0A1 - Temps #2
+    ControlBoardTemp,
+    RTDTemp1,
+    RTDTemp2,
+    RTDTemp3,
+
+    // 0x0A2 - Temps #3 & Torque Shudder
+    CoolantTemp,
+    HotSpotTemp,
+    MotorTemp,
+    TorqueShudder,
+
+    // 0x0A3
+    AnalogInput1,
+    AnalogInput2,
+    AnalogInput3,
+    AnalogInput4,
+    AnalogInput5,
+    AnalogInput6,
+    AnalogInput7,
+    AnalogInput8,
+
+    // 0x0A4 - Digital Input Status
+    ForwardSwitch,
+    ReverseSwitch,
+    BrakeSwitch,
+    RegenDisableSwitch,
+    IgnitionSwitch,
+    StartSwitch,
+    ValetMode,
+    DigitalInput8,
+
+    // 0x0A5 - Motor Position Information
+    MotorElectricalAngle,
+    MotorSpeed,
+    InverterFrequency,
+    DeltaResolverAngle,
+
+    // 0x0A6 - Current Information
+    PhaseACurrent,
+    PhaseBCurrent,
+    PhaseCCurrent,
+    DCBusCurrent
+};
